+++ src/types/ChainsIds.ts
@@ -1,7 +1,9 @@
 export enum ChainsIds {
   BSC = 'bsc',
+  POLYGON = 'polygon',
+  AVALANCHE = 'avax',
 }
 
-export type TChainsIds = ChainsIds.BSC;
+export type TChainsIds = ChainsIds.BSC | ChainsIds.POLYGON | ChainsIds.AVALANCHE;
 
 export const NonEvmChainsIds: ChainsIds[] = [];